+++ features/admin.ask-password-flow-builder.v1/providers/ask-password-flow-builder-provider.tsx
@@ -22,16 +22,20 @@ import useAuthenticationFlowBuilderCore from
 import AuthenticationFlowBuilderCoreProvider
     from "@wso2is/admin.flow-builder-core.v1/providers/authentication-flow-builder-core-provider";
 import {
-    GovernanceConnectorInterface,
-    GovernanceConnectorUtils,
-    ServerConfigurationsConstants,
-    UpdateGovernanceConnectorConfigInterface,
     getConnectorDetails,
     updateGovernanceConnector
-} from "@wso2is/admin.server-configurations.v1";
+} from "@wso2is/admin.server-configurations.v1/api/governance-connectors";
+import {
+    ServerConfigurationsConstants
+} from "@wso2is/admin.server-configurations.v1/constants/server-configurations-constants";
 import {
     AskPasswordFormUpdatableConfigsInterface
 } from "@wso2is/admin.server-configurations.v1/models/ask-password";
+import {
+    GovernanceConnectorInterface,
+    UpdateGovernanceConnectorConfigInterface
+} from "@wso2is/admin.server-configurations.v1/models/governance-connectors";
+import { GovernanceConnectorUtils } from "@wso2is/admin.server-configurations.v1/utils/governance-connector-utils";
 import { AlertLevels } from "@wso2is/core/models";
 import { addAlert } from "@wso2is/core/store";
 import { useReactFlow } from "@xyflow/react";