+++ features/admin.groups.v1/components/wizard/create-group-wizard.tsx
@@ -34,7 +34,7 @@ import {
     RolesV2Interface
 } from "@wso2is/admin.roles.v2/models/roles";
 import { UserBasicInterface } from "@wso2is/admin.users.v1/models/user";
-import { PRIMARY_USERSTORE } from "@wso2is/admin.userstores.v1/constants";
+import { PRIMARY_USERSTORE } from "@wso2is/admin.userstores.v1/constants/user-store-constants";
 import { isFeatureEnabled } from "@wso2is/core/helpers";
 import {
     AlertLevels,