+++ modules/i18n/src/translations/en-US/portals/console.ts
@@ -8568,7 +8568,6 @@ export const console: ConsoleNS = {
             },
             oidcScopesEdit: {
                 backButton: "Go back to Scopes",
-                subTitle: "Add or remove OIDC attributes of the scope",
                 title: "Edit scope: {{ name }}"
             },
             organizations: {